feat: use double to store trainings and predictions

+++ include/caches/wlc.h
@@ -318,7 +318,7 @@ public:
     uint32_t n_req;
     int64_t n_early_stop = -1;
     uint32_t n_logging_start0;
-    vector<float> trainings_and_predictions;
+    vector<double> trainings_and_predictions;
     vector<uint16_t> training_and_prediction_logic_timestamps;
     string task_id;
     string dburl;
@@ -488,7 +488,7 @@ public:
             uploader.close();
             uploader = bucket.open_upload_stream(task_id + ".trainings_and_predictions");
             for (auto &b: trainings_and_predictions)
-                uploader.write((uint8_t *) (&b), sizeof(float));
+                uploader.write((uint8_t *) (&b), sizeof(double));
             uploader.close();
             uploader = bucket.open_upload_stream(task_id + ".training_and_prediction_timestamps");
             for (auto &b: training_and_prediction_logic_timestamps)
